requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Bcrypt>=1.0.0
PyMySQL>=1.1.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
easyocr>=1.7.0
Pillow>=10.0.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
setuptools>=65.0.0
google-generativeai>=0.3.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
requests>=2.31.0
razorpay>=1.4.0
qrcode[pil]>=7.4.0